Modularize questions (#6)
Questions on applications and in reviews can now be changed without changes to the database structure. Currently, the questions are being pulled in from json files, but eventually there will be an admin page that allows for direct manipulation of the database.

I've also changed the name of the project, the corresponding folder, and therefore all of the imports.